refactor: 자동차 이름 검증 로직을 carNameValidator에서 Car로 이동
